abstracted exoplanets inspection to component
